Workaround for traffic receive stop response
Traffic receive stop response is the largest one 520 * numStreams, this
is causing issues with the serial agent, as anything > 800 bytes over
UART is only partially read by the serial agent and leads to "function
not defined" error in the WFA CA as the regex match fails.

As a workaround remove the advanced Wi-Fi feature data structures from
the DUT response union to decrease the overall size, with this the
structure size reduces from 520 to ~260, and this works well with serial
agent.

We still need to maintain the TLV numbering, so, replace the removed
functions with not defined.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>